Added 3 separate controllers for each resource: BooksController, MembersController, BookingsController
